+++ geonode_mapstore_client/client/js/components/DetailsPanel/DetailsPanel.jsx
@@ -27,7 +27,8 @@ import mapTypeHOC from '@mapstore/framework/components/map/enhancers/mapType';
 import AuthorInfo from '@js/components/AuthorInfo/AuthorInfo';
 import Loader from '@mapstore/framework/components/misc/Loader';
 import { getUserName } from '@js/utils/SearchUtils';
-import FitBounds from '@mapstore/framework/components/geostory/common/map/FitBounds';
+import ZoomTo from '@js/components/ZoomTo';
+import { boundsToExtentString } from '@js/utils/CoordinatesUtils';
 
 const Map = mapTypeHOC(BaseMap);
 Map.displayName = 'Map';
@@ -124,15 +125,12 @@ const DefinitionListContainer = ({itemslist}) => {
 const MapThumbnailView = ({ initialBbox, layers, onMapThumbnail, onClose, savingThumbnailMap } ) => {
 
     const [currentBbox, setCurrentBbox] = useState();
+    const { bounds, crs } = initialBbox;
+    const extent = boundsToExtentString(bounds, crs);
 
     function handleOnMapViewChanges(center, zoom, bbox) {
         setCurrentBbox(bbox);
     }
-
-    const { bounds, crs } = initialBbox;
-    const { minx, miny, maxx, maxy } = bounds;
-    const extent = [minx, miny, maxx, maxy];
-    const projection = crs;
 
     return (
         <div>
@@ -158,21 +156,15 @@ const MapThumbnailView = ({ initialBbox, layers, onMapThumbnail, onClose, saving
                         ...(layers ? layers : [])
                     ]}
                 >
-                    <FitBounds
-                        mapType="openlayers"
-                        active
-                        geometry={extent}
-                        duration={300}
-                        geometryProjection={projection}
-                    />
+                    <ZoomTo extent={extent} />
                 </Map>
                 {savingThumbnailMap && <div className="gn-details-thumbnail-loader">
                     <Loader size={50} />
                 </div>
                 }
             </div>
-            <div className="gn-detail-extent-action" >
-                <Button className="btn-primary" onClick={() => onMapThumbnail(currentBbox)} ><Message msgId={"gnhome.apply"} /></Button><Button onClick={() => onClose() }><i className="fa fa-close"/></Button></div>
+            <div className="gn-detail-extent-action">
+                <Button className="btn-primary" onClick={() => onMapThumbnail(currentBbox, onClose)} ><Message msgId={"gnhome.apply"} /></Button><Button onClick={() => onClose() }><i className="fa fa-close"/></Button></div>
         </div>
     );
 
@@ -532,16 +524,9 @@ function DetailsPanel({
 
                     <div className="gn-details-panel-content-text">
                         <div className="gn-details-panel-title" >
-                            <div className="gn-details-panel-title-flex">
-                                <div>
-                                    <span className="gn-details-panel-title-icon" ><FaIcon name={icon} /> </span> <EditTitle disabled={!activeEditMode} tagName="h1"  title={resource?.title} onEdit={editTitle} >
+                            <span className="gn-details-panel-title-icon" ><FaIcon name={icon} /> </span> <EditTitle disabled={!activeEditMode} tagName="h1"  title={resource?.title} onEdit={editTitle} >
 
-                                    </EditTitle>
-                                </div>
-                                <div>
-                                    <ResourceStatus resource={resource} />
-                                </div>
-                            </div>
+                            </EditTitle>
 
                             {
                                 <div className="gn-details-panel-tools">
@@ -586,6 +571,7 @@ function DetailsPanel({
 
 
                         </div>
+                        <ResourceStatus resource={resource} />
                         {<p className="gn-details-panel-meta-text">
                             {resource?.owner &&  <div>{resource?.owner.avatar &&
                             <img src={resource?.owner.avatar} alt={getUserName(resource?.owner)} className="gn-card-author-image" />
