+++ geonode_mapstore_client/client/js/components/FiltersForm/FilterByExtent.jsx
@@ -7,55 +7,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState, useRef } from 'react';
 import { FormGroup, Checkbox } from 'react-bootstrap';
 import BaseMap from '@mapstore/framework/components/map/BaseMap';
 import mapType from '@mapstore/framework/components/map/enhancers/mapType';
 import Message from '@mapstore/framework/components/I18N/Message';
-import { reprojectBbox } from '@mapstore/framework/utils/CoordinatesUtils';
 import {
     boundsToExtentString,
     getFeatureFromExtent
 } from '@js/utils/CoordinatesUtils';
+import ZoomTo from '@js/components/ZoomTo';
 
 const Map = mapType(BaseMap);
 Map.displayName = 'Map';
-
-function ZoomTo({
-    map,
-    extent
-}) {
-    const once = useRef();
-    useEffect(() => {
-        if (map && extent && !once.current) {
-            const [
-                aMinx, aMiny, aMaxx, aMaxy,
-                bMinx, bMiny, bMaxx, bMaxy
-            ] = extent.split(',');
-            const projection = map.getView().getProjection().getCode();
-            let bounds;
-            const aBounds = reprojectBbox([aMinx, aMiny, aMaxx, aMaxy], 'EPSG:4326', projection);
-            if (bMinx !== undefined && bMiny !== undefined && bMaxx !== undefined && bMaxy !== undefined) {
-                const bBounds = reprojectBbox([bMinx, bMiny, bMaxx, bMaxy], 'EPSG:4326', projection);
-                // if there is the second bbox we should shift the minimum x value to correctly center the view
-                // the x of the [A] bounds needs to be shifted by the width of the [B] bounds
-                const minx = aBounds[0] - (bBounds[2] - bBounds[0]);
-                bounds = [minx, aBounds[1], aBounds[2], aBounds[3]];
-            } else {
-                bounds = aBounds;
-            }
-            map.getView().fit(bounds, {
-                size: map.getSize(),
-                duration: 300,
-                nearest: true
-            });
-            // ensure to avoid other fit action by setting once to true
-            once.current = true;
-        }
-    }, [ extent ]);
-
-    return null;
-}
 
 function FilterByExtent({
     id,
