+++ geonode_mapstore_client/client/js/components/ZoomTo/index.js
@@ -1,0 +1,3 @@
+import ZoomTo from "./ZoomTo";
+
+export default ZoomTo;
